allow overriding the default monitor via env var
