+++ src/hupper/__init__.py
@@ -1,7 +1,7 @@
 # public api
 # flake8: noqa
 
-from .compat import is_watchdog_supported
+from .utils import is_watchdog_supported
 
 from .reloader import (
     start_reloader,
